fix yearn helper
